requirements.txt
python-dotenv
google-cloud-documentai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
# oauth2client  # Removido por obsoleto
google-cloud-aiplatform
python-telegram-bot
Pillow
opencv-python
